Flatten `vf2` module structure
The previous module structure was matching the exact import structure
that was expected to be merged into Rustworkx.  This hasn't come to
pass yet, and the imports are easy to modify if necessary in the future.
Flattening the structure now makes it easier to read the file and to
make changes to it.
