+++ crates/circuit/src/dag_circuit.rs
@@ -37,10 +37,9 @@ use crate::operations::{
 use crate::packed_instruction::{PackedInstruction, PackedOperation};
 use crate::parameter::parameter_expression::ParameterExpression;
 use crate::register_data::RegisterData;
-use crate::rustworkx_core_vnext::isomorphism;
 use crate::slice::PySequenceIndex;
 use crate::variable_mapper::VariableMapper;
-use crate::{imports, Clbit, Qubit, Stretch, TupleLikeArg, Var, VarsMode};
+use crate::{imports, vf2, Clbit, Qubit, Stretch, TupleLikeArg, Var, VarsMode};
 
 use hashbrown::{HashMap, HashSet};
 use indexmap::IndexMap;
@@ -2198,18 +2197,18 @@ impl DAGCircuit {
             }
         };
 
-        isomorphism::vf2::is_isomorphic(
+        vf2::is_isomorphic(
             &self.dag,
             &other.dag,
             node_match,
-            isomorphism::vf2::NoSemanticMatch,
+            vf2::NoSemanticMatch,
             true,
             Ordering::Equal,
             true,
             None,
         )
         .map_err(|e| match e {
-            isomorphism::vf2::IsIsomorphicError::NodeMatcherErr(e) => e,
+            vf2::IsIsomorphicError::NodeMatcherErr(e) => e,
             _ => {
                 unreachable!()
             }
